Move heatmap-visual.tsx to visualisations.
